# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Centre Tour Booking Calendar API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /referral-rewards/{id}/claim:
    post:
      operationId: claimReferralReward
      tags: [dashboard]
      summary: Mark a referral reward as claimed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated referral reward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralReward"
components:
  schemas:
...
    ReferralReward:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        referrerName:
          type: string
        referrerPhone:
          type: string
        referrerEmail:
          type: ["string", "null"]
        rewardStatus:
          type: string
          enum: [pending, eligible, claimed]
        rewardDetails:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - bookingId
        - referrerName
        - referrerPhone
        - rewardStatus
        - createdAt
        - updatedAt
...